move loadVideo to delayed
